+++ packages/video-segment-generator/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fixed a bug where non-zero originPositions would result in incorrect periodPositions.
+
 ## [0.1.1] - 2021-04-07
 ### Changed
 - Update the required version of appliance-core to 0.6.1