+++ packages/video-segment-generator/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update `@tvkitchen/base-classes` to version `2.0.0-alpha.1`.
+
 ### Fixed
 - Fixed a bug where non-zero originPositions would result in incorrect periodPositions.