+++ packages/video-segment-generator/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2021-05-13
 ### Changed
 - Update `@tvkitchen/base-classes` to version `2.0.0-alpha.1`.
 - Update to support `origin` payloads instead of `timestamp`.
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial implementation of the `VideoSegmentGenerator` appliance.
 
-[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-segment-generator@0.1.1...HEAD
-[0.1.1]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-video-segment-generator@0.1.0
+[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-segment-generator@0.2.0...HEAD
+[0.2.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-video-segment-generator@0.2.0
+[0.1.1]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-video-segment-generator@0.1.1
 [0.1.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-video-segment-generator@0.1.0